+++ src/export.ts
@@ -7,4 +7,6 @@
  * @license Apache-2.0
  */
 
+// oxlint-disable oxc/no-barrel-file
+
 export * from '@matterbridge/core';