+++ CHANGELOG.md
@@ -72,7 +72,7 @@ These classes will run as threads in the next releases:
 - [matterbridge]: Bump `oxfmt` to v.0.55.0.
 - [matterbridge]: Bump `vitest` to v.4.1.9.
 - [matterbridge]: Bump `@vitest/coverage-v8` to v.4.1.9.
-- [oxlint]: Update Oxlint configuration to v.1.0.9.
+- [oxlint]: Update Oxlint configuration to v.1.0.10.
 
 ### Fixed